agent-backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.1
asyncpg==0.29.0
apscheduler==3.10.4
httpx==0.27.0
resend==2.3.0
python-dotenv==1.0.1
